Add separate Maven module for running
