+++ hyperbeam-single-machine/src/main/java/com/github/nhirakawa/hyperbeam/RayTracerRunnerComponent.java
@@ -1,0 +1,11 @@
+package com.github.nhirakawa.hyperbeam;
+
+import javax.inject.Singleton;
+
+import dagger.Component;
+
+@Singleton
+@Component(modules = HyperbeamModule.class)
+public interface RayTracerRunnerComponent {
+  RayTracerRunner buildRayTracerRunner();
+}
