+++ pom.xml
@@ -14,6 +14,7 @@
   <packaging>pom</packaging>
   <modules>
     <module>hyperbeam-core</module>
+    <module>hyperbeam-single-machine</module>
   </modules>
 
   <properties>
